+++ src/main/java/de/uib/configed/gui/FDialogRemoteControl.java
@@ -163,7 +163,6 @@ public class FDialogRemoteControl extends FEditStringList {
 	}
 
 	// DocumentListener
-
 	@Override
 	public void changedUpdate(DocumentEvent e) {
 		saveEditedText();